build!: migrate release bundles to Rolldown
BREAKING CHANGE: browser bundles now target ES2018 instead of ES5.
