+++ .github/copilot-instructions.md
@@ -6,7 +6,7 @@ The rules below are a Copilot-facing subset of the load-bearing safety guarantee
 
 ## Setup safety
 
-- Use Node.js 20 or newer for v2 development and validation.
+- Use Node.js 20.19 or newer for v2 development and validation. Rolldown currently requires `^20.19.0 || >=22.12.0`; the published Axios runtime floor remains Node.js 20.0.
 - Install with `npm ci`; the repo's `.npmrc` sets `ignore-scripts=true`. Do not remove that flag. If husky hooks are needed after a fresh install, run `npm rebuild husky && npx husky` once.
 - Do not add new runtime dependencies without discussion. `package-lock.json` is verified by `lockfile-lint` for npm HTTPS hosts and integrity hashes.
 - Package, lockfile, and GitHub Actions update PRs are maintainer/bot-only; close these PRs from outside collaborators. Keep the 7-day Dependabot delay unless a critical vulnerability requires a maintainer-led manual update.
@@ -17,7 +17,7 @@ The rules below are a Copilot-facing subset of the load-bearing safety guarantee
 - `lib/adapters/` — I/O: `xhr.js`, `http.js`, `fetch.js`. Default preference `['xhr', 'http', 'fetch']`, picked by capability detection in `lib/adapters/adapters.js`. Never branch on environment name.
 - `lib/platform/` — Node by default; browser builds alias to `lib/platform/browser`.
 - `lib/helpers/` — generic, reusable utilities; no axios-specific lifecycle logic here.
-- Source is ESM (`type: module`) with explicit `.js` import extensions. `dist/` is generated by Rollup — never edit it by hand. Keep `index.d.ts` (ESM) and `index.d.cts` (CJS) in sync for any public API change.
+- Source is ESM (`type: module`) with explicit `.js` import extensions. `dist/` is generated by Rolldown — never edit it by hand. Browser artifacts target ES2018; UMD remains classic-script/AMD compatible and exposes `axios`, but v2 does not promise ES5 syntax or runtime polyfills. Keep `index.d.ts` (ESM) and `index.d.cts` (CJS) in sync for any public API change.
 
 ## Error handling
