+++ .github/workflows/release-branch.yml
@@ -42,6 +42,9 @@ jobs:
       - name: Build project
         run: npm run build
 
+      - name: Validate build artifacts
+        run: npm run test:build:artifacts
+
       - name: Install Playwright with deps
         run: npx playwright install --with-deps
