+++ AGENTS.md
@@ -6,7 +6,7 @@ This file is the canonical contributor guide for both human and AI agents workin
 
 ## Setup And Safety
 
-- Use Node.js 20 or newer for v2 development and validation.
+- Use Node.js 20.19 or newer for v2 development and validation. Rolldown currently requires `^20.19.0 || >=22.12.0`; the published Axios runtime floor remains Node.js 20.0.
 - Use `npm ci`; repo `.npmrc` sets `ignore-scripts=true`, and CI also uses `npm ci --ignore-scripts`.
 - Do not remove `ignore-scripts=true`; if git hooks are needed after a fresh install, run `npm rebuild husky && npx husky` once.
 - Adding or updating dependencies is security-sensitive; `package-lock.json` is checked by `lockfile-lint` for npm HTTPS hosts and integrity hashes.
@@ -16,7 +16,8 @@ This file is the canonical contributor guide for both human and AI agents workin
 
 ## Commands
 
-- Build published artifacts: `npm run build` (`gulp clear` deletes `dist/`, then Rollup writes browser ESM/UMD/CJS and Node CJS bundles).
+- Build published artifacts: `npm run build` (`gulp clear` deletes `dist/`, then Rolldown writes browser ESM/UMD/CJS and Node CJS bundles).
+- Validate generated package contracts after a build: `npm run test:build:artifacts` (file inventory, ES2018 browser syntax, UMD global/AMD loaders, ESM/CJS export shapes, source maps, and platform separation).
 - Lint source only: `npm run lint`; focused lint: `npx eslint lib/path/to/file.js`.
 - Unit tests: `npm run test:vitest:unit`; focused unit test: `npm run test:vitest:unit -- tests/unit/path.test.js`.
 - Browser tests need Playwright installed first (`npx playwright install` locally; CI uses `npx playwright install --with-deps`); run `npm run test:vitest:browser:headless` for CI parity.
@@ -26,7 +27,8 @@ This file is the canonical contributor guide for both human and AI agents workin
 ## Package Shape
 
 - Source is ESM (`type: module`); public ESM entry is `index.js`, which re-exports the default instance from `lib/axios.js`.
-- Do not edit `dist/` by hand; it is ignored and generated from `lib/` by Rollup.
+- Do not edit `dist/` by hand; it is ignored and generated from `lib/` by Rolldown.
+- Browser ESM, UMD, and CJS outputs target ES2018. The UMD files remain classic-script/AMD compatible and expose the `axios` global, but v2 does not promise ES5 syntax or runtime polyfills.
 - Runtime package exports are split by environment: browser/react-native map Node HTTP/platform files to browser/null replacements, while Node CJS ships as `dist/node/axios.cjs`.
 - Keep public runtime exports, `index.d.ts` (ESM types), and `index.d.cts` (CJS `export = axios` types) in sync for API changes.
 - `lib/env/data.js` is version-generated by `gulp version`; do not edit it for normal feature work.
@@ -41,7 +43,7 @@ This file is the canonical contributor guide for both human and AI agents workin
 
 - `lib/core/` is axios domain logic: request dispatch, config merge, interceptors, headers, errors. Key classes: `Axios` (request dispatch + interceptor chains), `AxiosError` (standardized error codes), `AxiosHeaders` (case-insensitive header normalization), `InterceptorManager` (sync/async interceptor registration).
 - `lib/adapters/` performs I/O; default adapter preference is `['xhr', 'http', 'fetch']`, with capability selection in `lib/adapters/adapters.js`. Detect by capability, not environment name.
-- `lib/platform/` selects Node by default; browser builds rely on package/rollup aliasing to `lib/platform/browser`.
+- `lib/platform/` selects Node by default; browser builds rely on package/Rolldown browser-field resolution to `lib/platform/browser`.
 - `lib/helpers/` should stay generic and reusable outside axios; do not put axios-specific request lifecycle logic there.
 - New `lib/**/*.js` files should match existing source style: ESM imports with explicit `.js` extensions, `'use strict';` where current library files use it, and `AxiosError` for axios-originated failures.
 
@@ -93,7 +95,7 @@ This file is the canonical contributor guide for both human and AI agents workin
 
 ## Tests
 
-- Test layout is runtime-first: `tests/unit/**/*.test.js`, `tests/browser/**/*.browser.test.js`, `tests/smoke/esm/**/*.smoke.test.js`, `tests/smoke/cjs/**/*.smoke.test.cjs`.
+- Test layout is runtime-first: `tests/unit/**/*.test.js`, `tests/build/**/*.test.js`, `tests/browser/**/*.browser.test.js`, `tests/smoke/esm/**/*.smoke.test.js`, `tests/smoke/cjs/**/*.smoke.test.cjs`.
 - Use `tests/setup/server.js` for local HTTP servers and cleanup with `try/finally`; leaking servers causes Vitest hangs.
 - Keep CJS and ESM smoke coverage aligned when behavior is packaging/import related.
 - Type compatibility is exercised against Node 20 definitions through `tests/module/cjs` with TypeScript 4.9 and `tests/module/esm` with TypeScript 5.x; run the matching module suite for declaration changes.
