+++ PRE_RELEASE_DOCS.md
@@ -20,6 +20,16 @@ Do not store raw diffs or line-number-only instructions here; prefer stable sect
 
 ## Unreleased
 
+### Axios v2 ES2018 browser bundles and retained UMD
+
+- **Change:** Document ES2018 as the Axios v2 browser-bundle syntax floor while retaining the existing UMD CDN and loader contract.
+- **Source:** `PRE_RELEASE_CHANGELOG.md` Breaking Changes, Browser bundle syntax baseline.
+- **Status:** Pending.
+- **Docs targets:** README browser support and CDN installation sections; the v2 migration guide; docs feature/support pages; classic-script and AMD examples; Spanish, French, and Chinese translations after the English wording is finalized.
+- **Required content:** State that browser ESM, browser CommonJS, and UMD bundles target ES2018. Keep the existing `dist/axios.js` and `dist/axios.min.js` paths, jsDelivr/unpkg defaults, classic `<script>` usage, global `axios`, and AMD/RequireJS behavior. Remove the “ES5 UMD” label and direct consumers that require ES5 parsing to Axios v1.x or a controlled downstream transpilation step. Explain that syntax lowering does not provide runtime polyfills for APIs such as `Promise`, `URL`, or `Symbol`. Do not present `<script type="module">` as a transparent replacement for UMD because module scripts have module scope, different loading/CORS behavior, and do not automatically create `window.axios`.
+- **Examples:** Keep the existing classic `<script src=".../dist/axios.min.js"></script>` and AMD examples, and verify they continue using the same global/default Axios instance. Add migration wording rather than replacing those examples with ESM.
+- **Notes:** Apply these updates during v2 release preparation, not to documentation presenting the current v1.x release. Treat a future package-manager CommonJS removal separately from the retained browser UMD delivery artifact, and update every maintained translation after the English wording is approved.
+
 ### Axios v2 Node.js 20+ baseline
 
 - **Change:** Document Node.js 20 as the minimum supported runtime for Axios v2.
