+++ package.json
@@ -119,12 +119,13 @@
     "dist/node/axios.cjs"
   ],
   "scripts": {
-    "build": "gulp clear && cross-env NODE_ENV=production rollup -c",
+    "build": "gulp clear && cross-env NODE_ENV=production rolldown -c",
     "version": "npm run build && git add package.json",
     "preversion": "gulp version",
     "test": "npm run test:vitest",
     "test:vitest": "vitest run",
     "test:vitest:unit": "vitest run --project unit",
+    "test:build:artifacts": "vitest run --config vitest.build.config.js",
     "test:vitest:browser": "vitest run --project browser",
     "test:vitest:browser:headless": "vitest run --project browser-headless",
     "test:vitest:watch": "vitest",
@@ -148,17 +149,9 @@
     "proxy-from-env": "^2.1.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.29.0",
-    "@babel/preset-env": "^7.29.5",
     "@commitlint/cli": "^21.2.0",
     "@commitlint/config-conventional": "^21.2.0",
     "@eslint/js": "^10.0.1",
-    "@rollup/plugin-alias": "^6.0.0",
-    "@rollup/plugin-babel": "^7.1.0",
-    "@rollup/plugin-commonjs": "^29.0.3",
-    "@rollup/plugin-json": "^6.1.0",
-    "@rollup/plugin-node-resolve": "^16.0.3",
-    "@rollup/plugin-terser": "^1.0.0",
     "@vitest/browser": "^4.1.9",
     "@vitest/browser-playwright": "^4.1.9",
     "abortcontroller-polyfill": "^1.7.8",
@@ -181,8 +174,7 @@
     "multer": "^2.2.0",
     "playwright": "^1.61.1",
     "prettier": "^3.9.4",
-    "rollup": "^4.60.4",
-    "rollup-plugin-bundle-size": "^1.0.3",
+    "rolldown": "1.2.4",
     "selfsigned": "^5.5.0",
     "stream-throttle": "^0.1.3",
     "typescript": "^5.9.3",
